+++ shop/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render, reverse, redirect, get_object_or_404
+from django.contrib.auth.decorators import login_required
 from django.contrib import messages
 from .models import ShopProducts
 from .forms import ProductForm
@@ -30,8 +31,14 @@ def product_info(request, product_id):
     return render(request, 'shop/product_info.html', context)
 
 
+@login_required
 def add_product(request):
     """ A view to return the admin page to add a new product """
+
+    # Only superuser can access this view
+    if not request.user.is_superuser:
+        msgs.error(request, 'This action is restricted to site admin users')
+        return redirect(reverse('home'))
 
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
@@ -54,7 +61,14 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
+    """ A view to update product details in database """
+
+    # Only superuser can access this view
+    if not request.user.is_superuser:
+        msgs.error(request, 'This action is restricted to site admin users')
+        return redirect(reverse('home'))
 
     product = get_object_or_404(ShopProducts, id=product_id)
     
@@ -80,7 +94,14 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
+    """ A view to remove a product from the database """
+
+    # Only superuser can access this view
+    if not request.user.is_superuser:
+        msgs.error(request, 'This action is restricted to site admin users')
+        return redirect(reverse('home'))
 
     product = get_object_or_404(ShopProducts, id=product_id)
     product.delete()
